Move GithubLoader._initGithubLoader logic into GithubLoader constructor
`GithubLoader::_initGithubLoader()` existed to defer initialization so
that the derived class's constructor could set properties that were
needed. Now that the values needed for initialization can be directly
passed to `super()`, the initialization of GithubLoader can occur
completely in the constructor. Also, some public members were removed
since they are not accessed externally and were only needed for
`_initGithubLoader()`.