[Feature] Support for enterprise Github repository
Add an extra parameters in the tsd.json file to support specifying the url to your github host.

Parameter name is githubHost and by default will point to github.com (so no change to current behavior).

If an enterprise github url is specified instead then the url for requests will be adapted to follow the scheme of enterprise github.