Automatically adding tsling ignore directive.
It's currently very hard to use **tsd** and **tslint** together, especially in automated builds with **Gulp** etc. as `tslint` lacks option to ignore certain files.

As implementing ignore option in **tslint** seems to stuck (palantir/tslint#73) I thought that maybe it could be handled on **tsd** side by automatically adding ignore header so the file is not linted.